pyproject.toml
[project]
name = "ai-research-assistance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.4.1",
    "docling>=2.71.0",
    "langchain>=1.2.7",
    "llama-index>=0.14.13",
    "openai>=2.16.0",
    "opencv-python>=4.13.0.90",
    "pillow>=11.3.0",
    "pymupdf>=1.26.7",
    "pypdf>=6.6.2",
    "streamlit>=1.53.1",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=4.57.6",
]
